requirements.txt
aiogram>=3.13.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
easyocr>=1.7.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
